+++ packages/amplify-graphql-directives/src/directives/appsync/aws-lambda.ts
@@ -1,0 +1,13 @@
+import { Directive } from '../directive';
+
+const name = 'aws_lambda';
+const definition = /* GraphQL */ `
+  directive @${name} on FIELD_DEFINITION | OBJECT
+`;
+const defaults = {};
+
+export const AwsLambdaDirective: Directive = {
+  name,
+  definition,
+  defaults,
+};