+++ packages/amplify-graphql-directives/src/directives/v1/index.ts
@@ -1,0 +1,9 @@
+export { AuthDirectiveV1 } from './auth';
+export { ConnectionDirectiveV1 } from './connection';
+export { FunctionDirectiveV1 } from './function';
+export { HttpDirectiveV1 } from './http';
+export { KeyDirectiveV1 } from './key';
+export { ModelDirectiveV1 } from './model';
+export { PredictionsDirectiveV1 } from './predictions';
+export { SearchableDirectiveV1 } from './searchable';
+export { VersionedDirectiveV1, VersionedDirectiveV1Defaults } from './versioned';