+++ packages/amplify-graphql-name-mapping-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.9](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@3.4.8...@aws-amplify/graphql-maps-to-transformer@3.4.9) (2024-03-28)
+
+**Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer
+
 ## [3.4.8](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@3.4.7...@aws-amplify/graphql-maps-to-transformer@3.4.8) (2024-03-13)
 
 **Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer