+++ packages/graphql-versioned-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.75](https://github.com/aws-amplify/amplify-category-api/compare/graphql-versioned-transformer@5.2.74...graphql-versioned-transformer@5.2.75) (2024-03-28)
+
+**Note:** Version bump only for package graphql-versioned-transformer
+
 ## [5.2.74](https://github.com/aws-amplify/amplify-category-api/compare/graphql-versioned-transformer@5.2.73...graphql-versioned-transformer@5.2.74) (2024-02-28)
 
 **Note:** Version bump only for package graphql-versioned-transformer